Use u32 instead of u16 for opcode flags
The "struct opcode" is 64-bit, because of alignment. That means that we
previously had 16 bits of padding after the flags field, that are just
wasted.

Use this wasted space by integrating it into the flags field, by
promoting it from u16 to u32.

+++ disassembler.c
@@ -128,7 +128,7 @@ static const char *opcode_multdiv_flags[] = {
 	"No div check",
 };
 
-static int print_flags(char *buf, size_t len, u16 flags,
+static int print_flags(char *buf, size_t len, u32 flags,
 		       const char **array, size_t array_size,
 		       bool is_io)
 {
