+++ lightrec-private.h
@@ -161,7 +161,7 @@ struct lightrec_state {
 };
 
 u32 lightrec_rw(struct lightrec_state *state, union code op,
-		u32 addr, u32 data, u16 *flags,
+		u32 addr, u32 data, u32 *flags,
 		struct block *block);
 
 void lightrec_free_block(struct lightrec_state *state, struct block *block);
